requirements.txt
flask
google-genai
python-dotenv
gunicorn
xhtml2pdf
python-bidi==0.4.2
stripe
Werkzeug